Add support for `access_token` query string as part of POP protocol
Access_token needs to be supported in order for the websocket connection
to success from a web browser. The API within web browsers cannot set
custom headers. So, this is the workaround provided for now until
the API for web browsers is updated somehow to allow for setting
bearer tokens in the auth headers.

See: https://tools.ietf.org/html/rfc6750#section-2.3